fix token, feat edit comentarios
