+++ client/src/components/login-singUp/sing-up.tsx
@@ -5,7 +5,7 @@ import { useAppDispatch, useAppSelector } from "../../app/hooks";
 
 import {
   User_Register_State,
-  auth_Login_Google,
+
   selectUser,
 } from "../../features/counter/counterSlice";
 import GoogleAuth from "../GoogleAuth/GoogleAuth";
